requirements.txt
python-telegram-bot[webhooks]==21.3
SQLAlchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
cloudinary==1.40.0
paypalrestsdk==1.13.3
httpx==0.27.0
python-dotenv==1.0.1
Pillow==10.3.0
apscheduler==3.10.4
